error handling stuff and rename stdio.rs
